# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Custom Prysm endpoints for beacon node operations, validator metrics, and extended functionality.\nThese endpoints extend the standard Ethereum Beacon Node API with Prysm-specific features.",
        "title": "Prysm Beacon Node Custom API",
        "contact": {
            "name": "Prysm Team",
            "url": "https://github.com/OffchainLabs/prysm"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/OffchainLabs/prysm/blob/develop/LICENSE.md"
        },
        "version": "1.0"
    },
    "host": "localhost:3500",
    "basePath": "/",
    "paths": {
...
        "/prysm/v1/beacon/blobs": {
            "post": {
                "description": "Publishes blob sidecars to the beacon network. Only supported for Deneb through Fulu epochs.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Beacon"
                ],
                "summary": "Publish blob sidecars (DEPRECATED post-Fulu)",
                "deprecated": true,
                "parameters": [
                    {
                        "description": "Blob sidecars to publish",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.PublishBlobsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blobs successfully published"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "503": {
                        "description": "Beacon node is currently syncing",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
...
        "/prysm/v1/beacon/chain_head": {
            "get": {
                "description": "Returns detailed information about the current head, finalized, and justified checkpoints of the beacon chain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Beacon"
                ],
                "summary": "Get chain head information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.ChainHead"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
        "/prysm/v1/beacon/individual_votes": {
            "post": {
                "description": "Returns detailed voting information for specified validators in a given epoch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Beacon"
                ],
                "summary": "Get individual validator votes",
                "parameters": [
                    {
                        "description": "Validator indices/public keys and epoch",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.GetIndividualVotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.GetIndividualVotesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
...
        "/prysm/v1/beacon/weak_subjectivity": {
            "get": {
                "description": "Computes the weak subjectivity checkpoint for safe checkpoint sync initialization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Beacon"
                ],
                "summary": "Get weak subjectivity checkpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.GetWeakSubjectivityResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "503": {
                        "description": "Beacon node is currently syncing",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputil.DefaultJsonError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "structs.BeaconBlockHeader": {
            "type": "object",
            "properties": {
                "body_root": {
                    "type": "string"
                },
                "parent_root": {
                    "type": "string"
                },
                "proposer_index": {
                    "type": "string"
                },
                "slot": {
                    "type": "string"
                },
                "state_root": {
                    "type": "string"
                }
            }
        },
        "structs.BlobSidecars": {
            "type": "object",
            "properties": {
                "sidecars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.Sidecar"
                    }
                }
            }
        },
        "structs.ChainHead": {
            "type": "object",
            "properties": {
                "finalized_block_root": {
                    "type": "string"
                },
                "finalized_epoch": {
                    "type": "string"
                },
                "finalized_slot": {
                    "type": "string"
                },
                "head_block_root": {
                    "type": "string"
                },
                "head_epoch": {
                    "type": "string"
                },
                "head_slot": {
                    "type": "string"
                },
                "justified_block_root": {
                    "type": "string"
                },
                "justified_epoch": {
                    "type": "string"
                },
                "justified_slot": {
                    "type": "string"
                },
                "optimistic_status": {
                    "type": "boolean"
                },
                "previous_justified_block_root": {
                    "type": "string"
                },
                "previous_justified_epoch": {
                    "type": "string"
                },
                "previous_justified_slot": {
                    "type": "string"
                }
            }
        },
        "structs.Checkpoint": {
            "type": "object",
            "properties": {
                "epoch": {
                    "type": "string"
                },
                "root": {
                    "type": "string"
                }
            }
        },
        "structs.GetIndividualVotesRequest": {
            "type": "object",
            "properties": {
                "epoch": {
                    "type": "string"
                },
                "indices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "public_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "structs.GetIndividualVotesResponse": {
            "type": "object",
            "properties": {
                "individual_votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.IndividualVote"
                    }
                }
            }
        },
...
        "structs.GetWeakSubjectivityResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/structs.WeakSubjectivityData"
                }
            }
        },
        "structs.IndividualVote": {
            "type": "object",
            "properties": {
                "current_epoch_effective_balance_gwei": {
                    "type": "string"
                },
                "epoch": {
                    "type": "string"
                },
                "inactivity_score": {
                    "type": "string"
                },
                "inclusion_distance": {
                    "type": "string"
                },
                "inclusion_slot": {
                    "type": "string"
                },
                "is_active_in_current_epoch": {
                    "type": "boolean"
                },
                "is_active_in_previous_epoch": {
                    "type": "boolean"
                },
                "is_current_epoch_attester": {
                    "type": "boolean"
                },
                "is_current_epoch_target_attester": {
                    "type": "boolean"
                },
                "is_previous_epoch_attester": {
                    "type": "boolean"
                },
                "is_previous_epoch_head_attester": {
                    "type": "boolean"
                },
                "is_previous_epoch_target_attester": {
                    "type": "boolean"
                },
                "is_slashed": {
                    "type": "boolean"
                },
                "is_withdrawable_in_current_epoch": {
                    "type": "boolean"
                },
                "public_keys": {
                    "type": "string"
                },
                "validator_index": {
                    "type": "string"
                }
            }
        },
...
        "structs.PublishBlobsRequest": {
            "type": "object",
            "properties": {
                "blob_sidecars": {
                    "$ref": "#/definitions/structs.BlobSidecars"
                },
                "block_root": {
                    "type": "string"
                }
            }
        },
...
        "structs.Sidecar": {
            "type": "object",
            "properties": {
                "blob": {
                    "type": "string"
                },
                "index": {
                    "type": "string"
                },
                "kzg_commitment": {
                    "type": "string"
                },
                "kzg_commitment_inclusion_proof": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kzg_proof": {
                    "type": "string"
                },
                "signed_block_header": {
                    "$ref": "#/definitions/structs.SignedBeaconBlockHeader"
                }
            }
        },
        "structs.SignedBeaconBlockHeader": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/structs.BeaconBlockHeader"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "structs.WeakSubjectivityData": {
            "type": "object",
            "properties": {
                "state_root": {
                    "type": "string"
                },
                "ws_checkpoint": {
                    "$ref": "#/definitions/structs.Checkpoint"
                }
            }
        }
    }
}